Fix: normalize SEO title in draft generation and avoid pre-filling seoTitle on idea create
- Move normalizeSeoTitle to utils and use it in both UI and aiService
- Apply normalization in generateDraftFromIdea to avoid duplicated titles
- Initialize seoTitle as empty on idea creation (kanbanService)
- Update tests and PR body
